fix: Promise chaining
Made changes to the EnhancedPromise class to enhance the display of the call stack and handle results more effectively. Now, the methods wrappedResolve and wrappedReject accept results and process them, allowing the execution of functions if the result is a function. This improves code flexibility and readability. Additionally, some minor optimizations have been introduced to enhance the code structure. EnhancedPromise.pendingPromises is now assigned to pendingPromises for a clearer module export structure.